pyproject.toml
[tool.poetry]
name = "faces"
version = "0.1.0"
description = ""
authors = ["Ruslan Sirazhetdinov <ruslansir@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dlib = "^19.22.0"
numpy = "^1.20.3"
opencv-python = "^4.5.2"
piexif = "^1.1.3"
Pillow = "^8.2.0"
ExifRead = "^2.3.2"
pyheif = "^0.5.1"
black = "^21.6b0"
isort = "^5.8.0"
flake8 = "^3.9.2"
mypy = "^0.902"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
six = "^1.16.0"
tinydb = "^4.5.0"
pydantic = "^1.8.2"
python-dotenv = "^0.18.0"
ffmpeg-python = "^0.2.0"
python-magic = "^0.4.24"
redis = "^3.5.3"
walrus = "^0.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
